CRUD operation endpoints addead for goal

+++ src/main/java/mif/vu/ikeea/Entity/Repository/GoalRepository.java
@@ -3,6 +3,6 @@ package mif.vu.ikeea.Entity.Repository;
 import mif.vu.ikeea.Entity.Goal;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GoalRepository extends CrudRepository<Goal, Integer> {
+public interface GoalRepository extends CrudRepository<Goal, Long> {
 
 }
